Separation of Suzipu plugin from application in progress
