+++ src/notes_to_image.py
@@ -6,10 +6,10 @@ import music21
 import argparse
 import json
 
-from src.auxiliary import _create_suzipu_images
-from src.modes import GongdiaoModeList
+from src.fingering import FingeringProperties, Fingering, fingering_to_lowest_note
+from src.plugins.suzipu_lvlvpu_gongchepu.common import GongcheMelodySymbol, GongdiaoModeList, SuzipuAdditionalSymbol, \
+    suzipu_to_info
 from src.config import JIANPU_IMAGE_PATH, FIVELINE_IMAGE_PATH, CHINESE_FONT_FILE, SUZIPU_NOTATION_IMAGE_PATH
-from src.suzipu import suzipu_to_info, SuzipuAdditionalSymbol, GongcheMelodySymbol
 
 accidental_dictionary = {
             "None": None,
@@ -76,7 +76,7 @@ class NotationResources:
 def parse_arguments():
     parser = argparse.ArgumentParser(description="WriteSuzipuListToFile")
     parser.add_argument("--music_list", required=True,
-                        help="JSON string containing the Suzipu information.")
+                        help="JSON string containing the notational information.")
     parser.add_argument("--lyrics_list", required=True,
                         help="JSON string containing the lyrics information.")
     parser.add_argument("--output_file_path", default="temp.png", required=False,
@@ -90,41 +90,6 @@ def parse_arguments():
     parser.add_argument("--preface", default="", required=False,
                         help="Piece preface.")
     return parser.parse_args()
-
-
-class FingeringProperties:
-    def __init__(self, name, chinese_name, transposition_index, image_data):
-        self.name = name
-        self.chinese_name = chinese_name
-        self.transposition_index = transposition_index
-        self.image_data = image_data
-
-
-@dataclasses.dataclass
-class Fingering:
-    ALL_CLOSED_AS_1: FingeringProperties = FingeringProperties(name="Lowest = •1", chinese_name="筒音作 •1", transposition_index="-P8", image_data=("1", "low", None))
-    ALL_CLOSED_AS_2: FingeringProperties = FingeringProperties(name="Lowest = •2", chinese_name="筒音作 •2", transposition_index="-m7", image_data=("2", "low", None))
-    ALL_CLOSED_AS_3: FingeringProperties = FingeringProperties(name="Lowest = •3", chinese_name="筒音作 •3", transposition_index="-m6", image_data=("3", "low", None))
-    ALL_CLOSED_AS_4: FingeringProperties = FingeringProperties(name="Lowest = •4", chinese_name="筒音作 •4", transposition_index="-P5", image_data=("4", "low", None))
-    ALL_CLOSED_AS_5: FingeringProperties = FingeringProperties(name="Lowest = •5", chinese_name="筒音作 •5", transposition_index="-P4", image_data=("5", "low", None))
-    ALL_CLOSED_AS_6: FingeringProperties = FingeringProperties(name="Lowest = •6", chinese_name="筒音作 •6", transposition_index="-m3", image_data=("6", "low", None))
-    ALL_CLOSED_AS_FLAT_7: FingeringProperties = FingeringProperties(name="Lowest = •♭7", chinese_name="筒音作 •♭7", transposition_index="-M2", image_data=("7", "low", "flat"))
-
-    @classmethod
-    def from_string(cls, string):
-        for fingering in dataclasses.astuple(cls()):
-            if string == fingering.name or string == fingering.chinese_name:
-                return fingering
-        fingering = Fingering.ALL_CLOSED_AS_1
-        print(f"Could not detect fingering from name '{string}'. Returned {fingering.name} instead.")
-        return fingering
-
-
-def fingering_to_lowest_note(fingering: FingeringProperties):
-    transposition = fingering.transposition_index
-    base_note = music21.note.Note("C4")
-    base_note = base_note.transpose(transposition)
-    return base_note
 
 
 def horizontal_composition(image_list: list):
