+++ src/widgets_annotation.py
@@ -1,14 +1,10 @@
-import dataclasses
 import tkinter as tk
 import tkinter.ttk
 
-import PIL
-from PIL import ImageTk
-
-from src.auxiliary import BoxProperty, open_file_as_tk_image, _create_suzipu_images
+from src.auxiliary import BoxProperty
 from src.config import CHINESE_FONT_FILE
 from src.intelligent_assistant import predict_from_images
-from src.suzipu import SuzipuMelodySymbol, suzipu_to_info, SuzipuAdditionalSymbol, Symbol
+import importlib
 
 
 def exec_quick_fill_window_text(annotation_type_var, max_length_var):
@@ -108,52 +104,6 @@ def exec_intelligent_fill_window_text(annotation_type_var, max_length_var, get_b
     return exit_save_var.get(), prediction.get()
 
 
-def exec_quick_fill_window_suzipu(annotation_type_var, max_length_var):
-    quick_fill_window = tk.Toplevel()
-
-    exit_save_var = tk.BooleanVar()
-    exit_save_var.set(False)
-    def on_destroy_save_changes():
-        exit_save_var.set(True)
-        quick_fill_window.destroy()
-
-    annotation_type = annotation_type_var.get()
-    max_length = int(max_length_var.get())
-
-    quick_fill_window.title(f"Quick Fill {annotation_type}")
-
-    text_variable = tk.StringVar(quick_fill_window)
-    text_variable.set("")
-
-    left_character_string = tk.StringVar(quick_fill_window)
-    left_character_string.set("")
-    ok_button = tk.Button(quick_fill_window, text="OK", command=on_destroy_save_changes, state="disabled")
-
-    def on_change_text(*args):
-        #if len(text_variable.get()) > 0:
-        #    text_variable.set(text_variable.get()[0:max_length])
-        cell_number = len(text_variable.get().split("|"))
-        left_character_string.set(f"{cell_number}/{max_length} cells (separator: '|')")
-
-        if cell_number == max_length:
-            ok_button.configure(state="normal")
-        else:
-            ok_button.configure(state="disabled")
-
-    text_variable.trace("w", on_change_text)
-
-    tk.Entry(quick_fill_window, width=min(max_length * 2, 20), font=CHINESE_FONT_FILE,
-             # use *2, because Chinese characters have double width
-             textvariable=text_variable).pack()
-    tk.Label(quick_fill_window, textvariable=left_character_string).pack()
-    ok_button.pack()
-    on_change_text()
-
-    quick_fill_window.wait_window()
-
-    return exit_save_var.get(), text_variable.get()
-
-
 class TextAnnotationFrame:
     def __init__(self, window_handle, current_text_annotation_variable, quick_fill_vars, on_fill_all_boxes_of_type, get_box_images):
         self.window_handle = window_handle
@@ -186,93 +136,6 @@ class TextAnnotationFrame:
 
         for widget in self._widgets:
             widget.pack(padx=10, pady=5)
-
-    def get_frame(self):
-        return self.frame
-
-    def set_state(self, boolean):
-        state = "disabled"
-        if boolean:
-            state = "normal"
-        for widget in self._widgets:
-            widget.config(state=state)
-
-
-class SuzipuAnnotationFrame:
-    def __init__(self, window_handle, first_musical_var, second_musical_var, on_change_annotation=lambda: None, quick_fill_vars=[], on_fill_all_boxes_of_type=lambda: None):
-        self.window_handle = window_handle
-        self.frame = tk.LabelFrame(self.window_handle, text="Suzipu Annotation")
-        self.first_musical_var = first_musical_var
-        self.second_musical_var = second_musical_var
-        self.on_change_annotation = on_change_annotation
-        self.quick_fill_vars = quick_fill_vars
-        self.on_fill_all_boxes_of_type = on_fill_all_boxes_of_type
-
-        self.display_first_symbol = None
-        self.display_second_symbol = None
-
-        self._widgets = []
-        self._button_images = _create_suzipu_images()
-
-        self._create_frame()
-
-    def update_musical_display_image(self):
-        self.display_first_symbol.config(image=self._button_images[str(self.first_musical_var.get())])
-        self.display_second_symbol.config(image=self._button_images[str(self.second_musical_var.get())])
-
-    def _create_frame(self):
-        current_suzipu_choice_frame = tk.Frame(self.frame)
-        self.display_first_symbol = tk.Label(current_suzipu_choice_frame,
-                                        image=self._button_images[Symbol.NONE],
-                                        relief="sunken", state="disabled")
-        self.display_second_symbol = tk.Label(current_suzipu_choice_frame,
-                                         image=self._button_images[Symbol.NONE],
-                                         relief="sunken", state="disabled")
-
-        def construct_suzipu_frame(frame, musical_var, prefix):
-            second_symbol_frame = tk.LabelFrame(frame, text=f"{prefix} Symbol")
-
-            none_button = tk.Radiobutton(second_symbol_frame, image=self._button_images[Symbol.NONE],
-                                         variable=musical_var,
-                                         value="None", indicator=0, state="disabled",
-                                         command=lambda: [self.on_change_annotation(), self.update_musical_display_image()])
-            none_button.grid(row=0, column=0)
-            self._widgets.append(none_button)
-
-            for idx, melody_var in enumerate(dataclasses.astuple(SuzipuMelodySymbol())):
-                current_button = tk.Radiobutton(second_symbol_frame, image=self._button_images[melody_var],
-                                                variable=musical_var,
-                                                value=melody_var, indicator=0, state="disabled",
-                                                command=lambda: [self.on_change_annotation(), self.update_musical_display_image()])
-                self._widgets.append(current_button)
-                current_button.grid(row=1, column=idx)
-            for idx, additional_var in enumerate(dataclasses.astuple(SuzipuAdditionalSymbol())):
-                current_button = tk.Radiobutton(second_symbol_frame, image=self._button_images[additional_var],
-                                                variable=musical_var,
-                                                value=additional_var, indicator=0, state="disabled",
-                                                command=lambda: [self.on_change_annotation(), self.update_musical_display_image()])
-                self._widgets.append(current_button)
-                current_button.grid(row=2, column=idx)
-            return second_symbol_frame
-
-        def on_quick_fill():
-            exit_save_var, text_variable = exec_quick_fill_window_suzipu(*self.quick_fill_vars)
-            if exit_save_var:
-                self.on_fill_all_boxes_of_type(text_variable)
-
-        quick_fill_button = tk.Button(self.frame, text="Quick Fill...", command=on_quick_fill, state="disabled")
-
-        self.display_first_symbol.grid(row=0, column=0, padx=10)
-        self.display_second_symbol.grid(row=1, column=0, padx=10)
-        self._widgets += [self.display_first_symbol, self.display_second_symbol, quick_fill_button]
-
-        symbol_frames = tk.Frame(self.frame)
-        construct_suzipu_frame(symbol_frames, self.first_musical_var, "First").grid(row=0, column=0, padx=10, pady=5)  # upper frame
-        construct_suzipu_frame(symbol_frames, self.second_musical_var, "Second").grid(row=1, column=0, padx=10, pady=5)  # lower frame
-        quick_fill_button.grid(row=2, column=0, padx=10, pady=10)
-
-        current_suzipu_choice_frame.grid(row=0, column=0)
-        symbol_frames.grid(row=0, column=1)
 
     def get_frame(self):
         return self.frame
@@ -329,7 +192,9 @@ class AnnotationFrame:
                  on_previous=lambda: None, on_next=lambda: None,
                  on_change_annotation=lambda: None,
                  on_fill_all_boxes_of_type=lambda: None,
-                 get_box_images=lambda: None):
+                 get_box_images=lambda: None,
+                 mode_variable=None,
+                 get_mode_string=lambda: None):
         self.window_handle = window_handle
         self.boxtype_variable = boxtype_variable
         self.display_variable = display_variable
@@ -343,6 +208,8 @@ class AnnotationFrame:
         self.on_change_annotation = lambda: [on_change_annotation(), self.update_musical_image_display()]
         self.on_fill_all_boxes_of_type = on_fill_all_boxes_of_type
         self.get_box_images = get_box_images
+        self.mode_variable = mode_variable
+        self.get_mode_string = get_mode_string
 
         self.frame = tk.LabelFrame(self.window_handle, text="Annotation")
         self.image = None
@@ -353,7 +220,7 @@ class AnnotationFrame:
                                                  self.on_next)
 
         self.text_annotation = None
-        self.suzipu_annotation = None
+        self.musical_annotation_frame = None
         self.box_excluded_checkbox = None
 
         self._create_frame()
@@ -366,30 +233,37 @@ class AnnotationFrame:
         self.selection_frame.set_state(boolean)
         if boolean:
             if self.boxtype_variable.get() == BoxProperty.MUSIC:
-                self.suzipu_annotation.set_state(True)
+                self.musical_annotation_frame.set_state(True)
                 self.text_annotation.set_state(False)
             else:
-                self.suzipu_annotation.set_state(False)
+                self.musical_annotation_frame.set_state(False)
                 self.text_annotation.set_state(True)
         else:
-            self.suzipu_annotation.set_state(False)
+            self.musical_annotation_frame.set_state(False)
             self.text_annotation.set_state(False)
 
     def set_image(self, image):
         self.current_box_image_display.configure(image=image)
 
     def update_musical_image_display(self):
-        self.suzipu_annotation.update_musical_display_image()
+        self.musical_annotation_frame.update_musical_display_image()
 
     def _create_frame(self):
+
+        module = importlib.import_module("src.plugins.suzipu")
+
         annotations_frame = tk.Frame(self.frame)
         self.text_annotation = TextAnnotationFrame(annotations_frame, self.current_annotation_text_variable,
                                                    [self.boxtype_variable, self.type_length_variable],
                                                    self.on_fill_all_boxes_of_type, self.get_box_images)
         self.text_annotation.get_frame().grid(row=0, column=0, padx=10, pady=20)
-        self.suzipu_annotation = SuzipuAnnotationFrame(annotations_frame, *self.musical_vars, self.on_change_annotation, [self.boxtype_variable, self.type_length_variable],
-                                                   self.on_fill_all_boxes_of_type)
-        self.suzipu_annotation.get_frame().grid(row=0, column=1, padx=10, pady=20)
+        music_frame = tk.LabelFrame(annotations_frame, text="Musical Annotation");
+        self.musical_annotation_frame = module.NotationAnnotationFrame(music_frame, *self.musical_vars,
+                                           self.on_change_annotation,
+                                           [self.boxtype_variable, self.type_length_variable],
+                                           self.on_fill_all_boxes_of_type, self.mode_variable, self.get_mode_string)
+        self.musical_annotation_frame.get_frame().pack()
+        music_frame.grid(row=0, column=1, padx=10, pady=20)
         self.box_excluded_checkbox = tk.Checkbutton(annotations_frame, text='Exclude this box from dataset',
                                                     variable=self.current_box_is_excluded, onvalue=True,
                                                     offvalue=False, command=self.on_change_annotation, state="disabled")
@@ -402,6 +276,12 @@ class AnnotationFrame:
 
         self.box_excluded_checkbox.grid(row=3, column=0, sticky="W")
         self.box_line_break_checkbox.grid(row=4, column=0, sticky="W")
+
+    def set_mode_properties(self, props: dict):
+        self.musical_annotation_frame.set_mode_properties(props)
+
+    def get_mode_properties(self):
+        return self.musical_annotation_frame.get_mode_properties()
 
     def get_frame(self):
         return self.frame
